Time: 1 ms (52.89%), Space: 8.9 MB (65.21%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.c
@@ -1,5 +1,5 @@
 bool isPalindrome(int x) {
-    int ori = x;
+    long long ori = x;
     long long rev = 0;
     if(x < 0)
         return false;
